# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Weekly Commitments API
  description: |
    API for the Weekly Commitments module – plan creation, lifecycle management,
    reconciliation, manager review, RCDO integration, and AI-assisted suggestions.
  version: 1.0.0
  contact:
    name: ST6 Engineering
servers:
  - url: /api/v1
    description: API v1 base path
security:
  - BearerAuth: []
...
paths:
...
  /weeks/{weekStart}/plans/{userId}:
    get:
      operationId: getUserPlan
      summary: Get a user's plan (manager-only)
      tags: [Plans]
      parameters:
        - $ref: '#/components/parameters/WeekStartPath'
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Plan found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /weeks/{weekStart}/plans/{userId}/commits:
    get:
      operationId: getUserPlanCommits
      summary: Get a user's commits for the selected week (manager-only)
      tags: [Plans, Commits]
      parameters:
        - $ref: '#/components/parameters/WeekStartPath'
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Commit list for the user's weekly plan
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeeklyCommit'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /weeks/{weekStart}/team/summary:
    get:
      operationId: getTeamSummary
      summary: Manager team dashboard
      description: |
        Returns a paginated, filtered summary of direct reports' plans
        for the given week.
      tags: [Manager]
      parameters:
        - $ref: '#/components/parameters/WeekStartPath'
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: state
          in: query
          schema:
            $ref: '#/components/schemas/PlanState'
        - name: outcomeId
          in: query
          schema:
            type: string
            format: uuid
        - name: incomplete
          in: query
          schema:
            type: boolean
        - name: nonStrategic
          in: query
          schema:
            type: boolean
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/ChessPriority'
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/CommitCategory'
      responses:
        '200':
          description: Team summary returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /weeks/{weekStart}/team/rcdo-rollup:
    get:
      operationId: getRcdoRollup
      summary: RCDO roll-up for manager dashboard
      description: |
        Returns commits grouped by RCDO outcome across direct reports,
        with chess priority distribution and non-strategic counts.
      tags: [Manager]
      parameters:
        - $ref: '#/components/parameters/WeekStartPath'
      responses:
        '200':
          description: RCDO roll-up returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RcdoRollupResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /rcdo/tree:
    get:
      operationId: getRcdoTree
      summary: Get RCDO hierarchy
      tags: [RCDO]
      responses:
        '200':
          description: RCDO tree returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RcdoTreeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /rcdo/search:
    get:
      operationId: searchRcdo
      summary: Typeahead search for RCDO outcomes
      tags: [RCDO]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RcdoSearchResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WeekStartPath:
      name: weekStart
      in: path
      required: true
      description: ISO date of the Monday that starts the week (e.g., 2026-03-09)
      schema:
        type: string
        format: date
...
  schemas:
    PlanState:
      type: string
      enum: [DRAFT, LOCKED, RECONCILING, RECONCILED, CARRY_FORWARD]
    ReviewStatus:
      type: string
      enum: [REVIEW_NOT_APPLICABLE, REVIEW_PENDING, CHANGES_REQUESTED, APPROVED]
    ChessPriority:
      type: string
      enum: [KING, QUEEN, ROOK, BISHOP, KNIGHT, PAWN]
    CompletionStatus:
      type: string
      enum: [DONE, PARTIALLY, NOT_DONE, DROPPED]
    LockType:
      type: string
      enum: [ON_TIME, LATE_LOCK]
    CommitCategory:
      type: string
      enum: [DELIVERY, OPERATIONS, CUSTOMER, GTM, PEOPLE, LEARNING, TECH_DEBT]
    WeeklyPlan:
      type: object
      required: [id, orgId, ownerUserId, weekStartDate, state, reviewStatus, version, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        ownerUserId:
          type: string
          format: uuid
        weekStartDate:
          type: string
          format: date
        state:
          $ref: '#/components/schemas/PlanState'
        reviewStatus:
          $ref: '#/components/schemas/ReviewStatus'
        lockType:
          $ref: '#/components/schemas/LockType'
        lockedAt:
          type: string
          format: date-time
          nullable: true
        carryForwardExecutedAt:
          type: string
          format: date-time
          nullable: true
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ValidationError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    WeeklyCommit:
      type: object
      required: [id, weeklyPlanId, title, description, expectedResult, progressNotes, tags, version, createdAt, updatedAt, validationErrors]
      properties:
        id:
          type: string
          format: uuid
        weeklyPlanId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        chessPriority:
          $ref: '#/components/schemas/ChessPriority'
        category:
          $ref: '#/components/schemas/CommitCategory'
        outcomeId:
          type: string
          format: uuid
          nullable: true
        nonStrategicReason:
          type: string
          nullable: true
        expectedResult:
          type: string
        confidence:
          type: number
          nullable: true
        tags:
          type: array
          items:
            type: string
        progressNotes:
          type: string
        estimatedHours:
          type: number
          minimum: 0
          maximum: 100
          nullable: true
        snapshotRallyCryId:
          type: string
          format: uuid
          nullable: true
        snapshotRallyCryName:
          type: string
          nullable: true
        snapshotObjectiveId:
          type: string
          format: uuid
          nullable: true
        snapshotObjectiveName:
          type: string
          nullable: true
        snapshotOutcomeId:
          type: string
          format: uuid
          nullable: true
        snapshotOutcomeName:
          type: string
          nullable: true
        carriedFromCommitId:
          type: string
          format: uuid
          nullable: true
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        validationErrors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        actual:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/WeeklyCommitActual'
    WeeklyCommitActual:
      type: object
      required: [commitId, actualResult, completionStatus]
      properties:
        commitId:
          type: string
          format: uuid
        actualResult:
          type: string
        completionStatus:
          $ref: '#/components/schemas/CompletionStatus'
        deltaReason:
          type: string
          nullable: true
        timeSpent:
          type: number
          nullable: true
...
    TeamSummaryResponse:
      type: object
      required: [weekStart, users, reviewStatusCounts, page, size, totalElements, totalPages]
      properties:
        weekStart:
          type: string
          format: date
        users:
          type: array
          items:
            $ref: '#/components/schemas/TeamMemberSummary'
        reviewStatusCounts:
          $ref: '#/components/schemas/ReviewStatusCounts'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    TeamMemberSummary:
      type: object
      required: [userId, commitCount, incompleteCount, issueCount, nonStrategicCount, kingCount, queenCount, isStale, isLateLock]
      properties:
        userId:
          type: string
          format: uuid
        displayName:
          type: string
          nullable: true
        planId:
          type: string
          format: uuid
          nullable: true
        state:
          $ref: '#/components/schemas/PlanState'
        reviewStatus:
          $ref: '#/components/schemas/ReviewStatus'
        commitCount:
          type: integer
        incompleteCount:
          type: integer
          description: Count of saved actuals with completionStatus != DONE for RECONCILED/CARRY_FORWARD plans
        issueCount:
          type: integer
          description: Count of commits with validation errors (missing required fields)
        nonStrategicCount:
          type: integer
        kingCount:
          type: integer
        queenCount:
          type: integer
        lastUpdated:
          type: string
          format: date-time
          nullable: true
        isStale:
          type: boolean
        isLateLock:
          type: boolean
    ReviewStatusCounts:
      type: object
      required: [pending, approved, changesRequested]
      properties:
        pending:
          type: integer
        approved:
          type: integer
        changesRequested:
          type: integer
    RcdoRollupResponse:
      type: object
      required: [weekStart, items, nonStrategicCount]
      properties:
        weekStart:
          type: string
          format: date
        items:
          type: array
          items:
            $ref: '#/components/schemas/RcdoRollupItem'
        nonStrategicCount:
          type: integer
    RcdoRollupItem:
      type: object
      required: [outcomeId, commitCount, kingCount, queenCount, rookCount, bishopCount, knightCount, pawnCount]
      properties:
        outcomeId:
          type: string
          format: uuid
        outcomeName:
          type: string
          nullable: true
        objectiveId:
          type: string
          format: uuid
          nullable: true
        objectiveName:
          type: string
          nullable: true
        rallyCryId:
          type: string
          format: uuid
          nullable: true
        rallyCryName:
          type: string
          nullable: true
        commitCount:
          type: integer
        kingCount:
          type: integer
        queenCount:
          type: integer
        rookCount:
          type: integer
        bishopCount:
          type: integer
        knightCount:
          type: integer
        pawnCount:
          type: integer
...
    RcdoTreeResponse:
      type: object
      required: [rallyCries]
      properties:
        rallyCries:
          type: array
          items:
            $ref: '#/components/schemas/RcdoCry'
    RcdoCry:
      type: object
      required: [id, name, objectives]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        objectives:
          type: array
          items:
            $ref: '#/components/schemas/RcdoObjective'
    RcdoObjective:
      type: object
      required: [id, name, rallyCryId, outcomes]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        rallyCryId:
          type: string
          format: uuid
        outcomes:
          type: array
          items:
            $ref: '#/components/schemas/RcdoOutcome'
    RcdoOutcome:
      type: object
      required: [id, name, objectiveId]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        objectiveId:
          type: string
          format: uuid
    RcdoSearchResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/RcdoSearchResult'
    RcdoSearchResult:
      type: object
      required: [id, name, objectiveId, objectiveName, rallyCryId, rallyCryName]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        objectiveId:
          type: string
          format: uuid
        objectiveName:
          type: string
        rallyCryId:
          type: string
          format: uuid
        rallyCryName:
          type: string
...
    ApiError:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Valid JWT but insufficient access
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ApiError'
...
    ServiceUnavailable:
      description: Backing service unavailable
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ApiError'